refactor: home으로 묶은 폴더 삭제
