--- src/app/(home)/page.tsx
@@ -1,5 +1,0 @@
-const Page = () => {
-  return <div>admin page</div>;
-};
-
-export default Page;
