+++ src/app/artists/table.type.tsx
