+++ src/app/coupons/table.type.tsx
