+++ src/app/events/[eventId]/dates/[dailyEventId]/routes/[shuttleRouteId]/buses/new/page.tsx
@@ -3,7 +3,7 @@
 import { useCallback } from 'react';
 import { useRouter } from 'next/navigation';
 import { Controller, useForm } from 'react-hook-form';
-import { conform, type CreateBusFormType } from './types/form.type';
+import { conform, type CreateBusFormType } from './form.type';
 import Input from '@/components/input/Input';
 import { Field, Label, Radio, RadioGroup } from '@headlessui/react';
 import { CheckIcon } from 'lucide-react';
