+++ src/app/events/[eventId]/dates/[dailyEventId]/routes/[shuttleRouteId]/table.type.tsx
