+++ src/app/events/[eventId]/dates/[dailyEventId]/table.type.tsx
