+++ src/app/hubs/new/page.tsx
@@ -3,7 +3,7 @@
 import { useRouter } from 'next/navigation';
 import { useCallback, useMemo } from 'react';
 import { Controller, useForm, useWatch } from 'react-hook-form';
-import { conform, type CreateHubFormType } from './types/form.type';
+import { conform, type CreateHubFormType } from './form.type';
 import Input from '@/components/input/Input';
 import CoordInput from '@/components/input/CoordInput';
 import RegionInput from '@/components/input/RegionInput';
