+++ src/app/hubs/page.tsx
@@ -6,7 +6,7 @@ import { useMemo } from 'react';
 import BlueLink from '@/components/link/BlueLink';
 import Heading from '@/components/text/Heading';
 import { useGetRegionHubs } from '@/services/location.service';
-import { columns } from './types/table.type';
+import { columns } from './table.type';
 import { PAGINATION_LIMIT } from '@/constants/config';
 import RegionHubFilter from './components/RegionHubFilter';
 import useRegionHubFilter from './hooks/useRegionHubFilter';
