+++ src/app/page.tsx
@@ -1,16 +1,10 @@
-import type { ReactNode } from 'react';
-
-interface Props {
-  children: ReactNode;
-}
-
-const Layout = async ({ children }: Props) => {
+const Page = () => {
   return (
     <main className="flex h-full w-full flex-col items-center justify-start gap-16 bg-white p-16">
       <h1 className="text-[32px] font-500">핸디버스 관리자 대시보드</h1>
-      {children}
+      <div>admin page</div>;
     </main>
   );
 };
 
-export default Layout;
+export default Page;
