--- src/app/reservations/(home)/loading.tsx
@@ -1,5 +1,0 @@
-const Loading = () => {
-  return <div>Loading...</div>;
-};
-
-export default Loading;
