+++ src/app/reservations/[route_id]/table.type.tsx
